+++ src/main/java/roomescape/dto/member/MemberRequest.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.dto.member;
 
 import roomescape.domain.member.MemberName;
